+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added `with_capacity` to `Dataset`
 - Reduced dependencies
 
 ## 0.1.0 (2021-12-01)